BDOG-399 extract common base of slugInfos repositories
